Moved Node and related types into own file
